python/requirements.txt
confluent-kafka==1.0.0
kafka-python==1.4.6
Flask==1.0
